+++ app/schemas/wish.py
@@ -1,4 +1,5 @@
 from typing import Optional
+from decimal import Decimal
 
 from pydantic import BaseModel
 
@@ -6,7 +7,7 @@ from pydantic import BaseModel
 class WishBase(BaseModel):
     title: str
     link: Optional[str] = None
-    price_estimate: Optional[float] = None
+    price_estimate: Optional[Decimal] = None
     notes: Optional[str] = None
